added undiagonalizermatrices[]
When diagonalizing the R-matrices, it can happen that some of the changed F-symbols can not be written in terms of the general number field that is used to represent the F- and R-symbols. Therefore, the command to `undiagonalize' the R-matrices was added. In addition, when obtaining the exact form of the F- and R-symbols, it is checked that one works with the original F- and R-symbols.

Some minor bugfixes.