saving now possible, some updates and bugfixes
The possibility to save and load data was added.

The exact form of the s-matrix elements was not general enough, so the routine to find the exact form of the modular data was changed accordingly.
This also means that the routine to check the exact modular data algebraically, does not work properly anymore. So, the algebraic checks of the data was disabled, thought the code remains in the package (so one can run these commands, at ones own risk, by putting alatc`Private` in front of them).

Some minor changes and bugfixes.